Update style.css/index.html
Updated style.css for lines 1-8 added a body style, lines 10-25 added a header stlye with a style for the header tag so that it displays to the right with a linear gradient style, lines 27-47 added a style for the aside so it displays on the left with the search for a city being bold and larger in font then the buttons in a column for a quickview of those cities weather, lines  49-54 added a style to the section so it aligns in the middle of the page with margin space between the 5-day forecast, lines 56-61 added styles for the main 5-day forecast so that it aligns properly with the searched for city. Updated index.html to have a weather emoji from windows, added a class to the label to adjust font/style of font, and added a main section to clarify the area of the html.