Update logic.js/index.html
Updated logic.js to include a fetch for the queryURL which I added so that I can target the "city" and then it will use the APIKey for the openweather api to give me a response then return the response then give me data however I was not able to get anydata. Updated the index.html added the links to the jquery themes, and jquery ui as well as the script links at the bottom of the body. Also added to the index.html some windows icons for the temp wind and humidity.