Update logic.js/index.html
Updated the logic.js file to have an event listener for the search button that grabs the id, prevents default form behavior for that id when submitted/clicked, takes the city input for the api function line 34, trims the value to display it properly, if there is no data in the search bar when clicked then alert to do so 23-30, then the function to get the weather data is set up to use the apikey and the url parameters will take in the city and use the apikey, then the data is fetched and returned as an object in the console. Commented in the other functions so that I can put them into the next function for the second fetch using the 5 day forecast api then locally stored the city name, and set the current day data to the date of the the data. Then started making the function to display the weather data for the 5day forecast however I dont have an object stored locally to work with yet so I can set the data to the paragraph tags for the 5 day forecast. Made a function to save cities and push the cities if there is one to the local storage. then made a function to display the saved cities by parseing the data from the local storage by the array. Updated the index.html so that there is ids on the cards for the 5 day data so I can target them in the javascript to populate as well as the section for the current weather now there is an id there to for the daily forecast.