Update logic.js/index.html
Updated the logic.js to have the elements for the wind and humidity captured for display on the 5 day forecast, however it does not populate the cards yet. Updated the index.html to have less buttons as I would want the buttons to pop up when the search is made for a search history section.