Update logic.js/index.html
Updated logic.js to have a the apikey at the top for global use, updated the weatherdata name to displayFiveDayData line 46, updated line 51-52 to make sure the data is displayed when the api is fetched, lines 54-55 set the current data to local storage, and the city name for saved city search history, updated lines 63-75 to have the 5 day forecast data populate by grabbing the id and setting the text content for the data, line 90-103 I made it so that when you type in a city there is a button that is generated for you so you can see what cities youve searched and then you can go back to them to compare, lines 105-133 is for grabbing the daily data for that days data however im not sure if ill need this function, lines 136-146 I added a function to populate the daily display with data for the daily forecast. Updated the index.html to have temp wind and humidity icons for the daily forecast display, updated lines 32-33 so that I can populate the search history under the search for city box with interactive search history, updated the h5 tags for the card title so that I can grab them and loop them through to populate, also added an icon id to grab onto so I can loop through the images for each day.