Updated logic.js to include a fivedayforecast that gathers the data for the fiveday forecast, changed index.html id attribute name
Updated the logic.js to include different parameters for the document.getElementById methods lines 159-162 however that doesnt display the fiveday forecast, and commented out lines 138-143 because it gave me an error that was not needed. Updated the index.html to just change the id of humiditynumber to humiditydaily to make more sense.