Updated logic.js to display the five day forecast.
Updated the logic.js to include const variables for the forecast so that i can extract the relevant data to set. Also included a displaySavedCities call at the bottom of the page so that the data for the weather can be refreshed and removed and then the user can either type in a new city or click on their search history to display past searches.